remove lua and value from shared crate
