+++ lib/shared/Cargo.toml
@@ -6,18 +6,10 @@ edition = "2018"
 publish = false
 
 [dependencies]
-bytes = { version = "0.5.6", optional = true }
-chrono = { version = "0.4", optional = true }
 metrics = { version = "0.13.0-alpha.8", optional = true }
 metrics-tracing-context = { version = "0.1.0-alpha.3", optional = true }
 nom = { version = "6", optional = true }
-paste = { version = "1", optional = true }
-remap = { package = "remap-lang", path = "../remap-lang", optional = true }
-rlua = { git = "https://github.com/kyren/rlua", rev = "25bd7e6bffef9597466a98bfca80a3056c9e6320", optional = true }
-serde = { version = "1", optional = true, default-features = false }
-serde_json = { version = "1", optional = true }
 snafu = { version = "0.6", optional = true }
-toml = { version = "0.5", optional = true, default-features = false }
 tracing = { version = "0.1", optional = true }
 tracing-futures = { version = "0.2", features = ["futures-01", "futures-03"], optional = true }
 tracing-limit = { path = "../tracing-limit", optional = true }
@@ -32,27 +24,13 @@ atty = { version = "0.2", optional = true }
 
 default = [
     "conversion",
-    "value",
     "trace",
     "tokenize",
     "test_util",
-    "lua",
 ]
 
 conversion = [
     "snafu",
-    "value",
-]
-
-value = [
-    "bytes",
-    "chrono",
-    "paste",
-    "remap",
-    "serde",
-    "serde_json",
-    "test_util",
-    "toml",
 ]
 
 trace = [
@@ -73,10 +51,4 @@ tokenize = [
 test_util = [
     "atty",
     "trace",
-]
-
-lua = [
-    "chrono",
-    "rlua",
-    "value",
 ]
