+++ lib/shared/src/lib.rs
@@ -1,9 +1,6 @@
 #[cfg(feature = "conversion")]
 pub mod conversion;
 
-#[cfg(feature = "value")]
-pub mod value;
-
 #[cfg(feature = "trace")]
 pub mod trace;
 
@@ -12,6 +9,3 @@ pub mod test_util;
 
 #[cfg(feature = "tokenize")]
 pub mod tokenize;
-
-#[cfg(feature = "lua")]
-pub mod lua;
