+++ scripts/check-style.sh
@@ -33,7 +33,6 @@ for FILE in $(git ls-files); do
     *ico) continue;;
     *sig) continue;;
     tests/data*) continue;;
-    lib/shared/tests/data*) continue;;
     distribution/kubernetes/*/*.yaml) continue;;
   esac
