+++ tests/data/fixtures/value/bytes/cn.json
