+++ tests/data/fixtures/value/bytes/emoji.json
