+++ tests/data/fixtures/value/bytes/friendly.json
