+++ tests/data/fixtures/value/bytes/jp-emoji.json
