+++ tests/data/fixtures/value/bytes/quotes.json
