+++ tests/data/fixtures/value/bytes/rtl.json
