+++ tests/data/fixtures/value/integer/maxint_32.json
