+++ tests/data/fixtures/value/integer/maxint_64.json
